% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUtils.R
\name{getGGAR}
\alias{getGGAR}
\title{Aspect ratio for a ggplot}
\usage{
getGGAR(p)
}
\arguments{
\item{p}{a ggplot object}
}
\description{
This internal function attempts to get the aspect ratio of a ggplot.
It takes multiple facets into account but only really works for plots where ggplot::coord_fixed() so that the
scales on the x and y axes are compatible.
}
\keyword{internal}
