% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUtils.R
\name{magicPlot}
\alias{magicPlot}
\title{Conveniently save a plot to a file}
\usage{
magicPlot(p, filename, type = "png", height = NULL, width = NULL, ...)
}
\arguments{
\item{p}{The plot to be saved.}

\item{filename}{The filename of the saved plot (character string) omitting the file type extension}

\item{type}{The type of file to make, defaults to "png" but also "pdf" and others are possible (see options to Cairo::Cairo function)}

\item{height, width}{The height and width of the plot (as numeric) units will depend on the type parameter,
it is pixels for "png" and inches for "pdf}

\item{...}{Passed to Cairo function}
}
\value{
Nothing, function prints a plot to a file.
}
\description{
Basically a convenience wrapper that opens graphical device (with Cairo), saves the provided ggplot object and closes it again.  Helpfully it will
automatically calculate and use an aspect ratio if possible (and not provided).
}
