% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotUtils.R
\name{nOptCols}
\alias{nOptCols}
\title{Get}
\usage{
nOptCols(x, rows.more.than.cols = 0, type = "Spatial", nlayers, ntimes)
}
\arguments{
\item{x}{Either the number of panels or something from which the number of panels can be derived
currently possible options are: a simple numeric (given the number of panels) or a Field (from
from which the number of panels are derived from the layers and additional possible facets)}

\item{rows.more.than.cols}{Preferred number of rows more than columns as an integer (can be negative).  For
example a value of "1" will optimise for more row than columns, "2" for two more rows, "0" for the same
number of rows as columns, "-1" for one less row than columns.  Note, the function doesn't guarantee
that it will work out like this, just that it will try to respect this preference if there is a choice.}

\item{type}{The type of plot as a character string (used for determing panels from dimensions).  Currently only
implemented for "Spatial".}

\item{nlayers}{Numeric for number of layers (in case only subset a subset of layers in x are plotted)}

\item{ntimes}{Numeric for number of time periods (Years and/or Days/Months/Seasons) (in case only subset a subset of
time periods in x are plotted)}
}
\description{
Get
}
